requirements.txt
asgiref==3.8.1
Django==5.0.12
django-extensions==3.2.3
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
django-cors-headers==4.7.0
psycopg2==2.9.10
djangorestframework-simplejwt==5.5.0
channels==4.2.2
channels_redis==4.2.1
daphne==4.1.2
mutagen==1.47.0
moviepy==2.1.2
django-filter==25.1
mutagen==1.47.0
